Scrolling modifiers: Support zooming (#337)
Zooming in and out is implemented by pressing <kbd>⌘+</kbd> and
<kbd>⌘-</kbd>.

An alternative way is to simulate a pinch gesture, but the effect is not
the same. This way would cause the horizontal scrollbar to appear in
apps like web browsers. We may provide this type of zooming as a
“gesture zoom” or something in the future.

Closes #227.