Scrolling modifiers: Add "Default Action", "Ignore Modifier" and update the behavior of "No Action" (#550)
In previous versions, the scrolling modifier "No Action" was actually
behaving as "Default Action". This patch addresses this issue and
corrects the behavior.

This change is not considered a breaking change because the modifier
action will automatically switch to "Default Action" if "No Action" was
used previously.

Additionally, two new modifier actions have been added: "Default Action"
and "Ignore Modifier". "Ignore Modifier" behaves as if the modifier key
was not pressed.

Closes #544.