Mouse move pointer events can be forwarded to the scroll wheel (#929)
I have two trackballs. I wanted one to behave as normal for mouse
movement and clicking, and the other to act as a dedicated scroll
device. I could not find any way to do this with any existing software,
so I forked this project and added a "Pointer redirects to scroll"
option under the pointer configuration settings.

I have extensively tested this functionality and it does exactly what I
needed, so I thought I would open a pull request for it to see if there
is interest in having this added to the project.